refactor: move files to new features/ folder structure
